R-MVP1-2: generalize the finalize helper to a shared .sergeant/lib location
CONFIRMED invariants:MVP1-R1-E2 / test-honesty:MVP1-TH-02 — R-MVP1-2's
actual ruling (meaning (3): a deterministic shared finalize helper any
workflow's closing stage can invoke) was never built; only the workflow-
local repo-to-icm/scripts/finalize.py existed, and the ruling's own L7 pin
had no test.

Move the generalized, workflow-agnostic implementation to
.sergeant/lib/finalize.py (WORKFLOW_ROOT is now a required argument — the
shared copy has no workflow of its own to default to). repo-to-icm's own
scripts/finalize.py becomes a thin backward-compatible wrapper that
forwards to the shared helper with its own root pinned, so 90-reconcile's
existing instructions and scripts/test-finalize-evidence-guard.py need no
change (verified: both still pass unchanged against the wrapper).

Add .sergeant/lib/test-finalize-disposition.py, running R-MVP1-2's pin
literally: a two-stage fixture (one evidence, one promote output) proves
the promoted file survives on the retained branch with the evidence file's
removal recoverable from branch history when the closing-stage instruction
runs, and that reverting the closing-stage instruction (never invoking
finalize) leaves the evidence file untouched.

Retrofitting every other admitted workflow's own closing stage to invoke
the shared helper is workflow content (R-MVP1-2's own boundary: "Owner:
workflow content") left as follow-up, not core's remaining scope here —
recorded in GAUNTLET.md.
